+++ darts/models/forecasting/tft_model.py
@@ -6,7 +6,6 @@ Temporal Fusion Transformer (TFT)
 from typing import Union, List, Optional, Tuple, Dict, Sequence
 
 import numpy as np
-from numpy.random import RandomState
 
 import torch
 from torch import nn
@@ -21,9 +20,9 @@ from darts.utils.data import (
     MixedCovariatesTrainingDataset,
     MixedCovariatesInferenceDataset,
 )
+from darts.models.forecasting.torch_pl_module import PLTorchParametricProbabilisticForecastingModel, PLMixedCovariatesTorchModel
 from darts.models.forecasting.torch_forecasting_model import (
     MixedCovariatesTorchModel,
-    TorchParametricProbabilisticForecastingModel
 )
 from darts.models.forecasting.tft_submodels import (
     _GateAddNorm,
@@ -38,8 +37,7 @@ logger = get_logger(__name__)
 MixedCovariatesTrainTensorType = Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]
 
 
-class _TFTModule(nn.Module):
-
+class _TFTModule(PLTorchParametricProbabilisticForecastingModel, PLMixedCovariatesTorchModel):
     def __init__(self,
                  output_dim: Tuple[int, int],
                  input_chunk_length: int,
@@ -52,7 +50,8 @@ class _TFTModule(nn.Module):
                  hidden_continuous_size: int = 8,
                  dropout: float = 0.1,
                  add_relative_index: bool = False,
-                 likelihood: Optional[Likelihood] = None):
+                 likelihood: Optional[Likelihood] = None,
+                 **kwargs):
 
         """ PyTorch module implementing the TFT architecture from `this paper <https://arxiv.org/pdf/1912.09363.pdf>`_
         The implementation is built upon `pytorch-forecasting's TemporalFusionTransformer
@@ -92,7 +91,7 @@ class _TFTModule(nn.Module):
             a ``QuantileRegression`` likelihood.
         """
 
-        super(_TFTModule, self).__init__()
+        super(_TFTModule, self).__init__(likelihood=likelihood, **kwargs)
 
         self.n_targets, self.loss_size = output_dim
         self.input_chunk_length = input_chunk_length
@@ -104,7 +103,6 @@ class _TFTModule(nn.Module):
         self.num_attention_heads = num_attention_heads
         self.full_attention = full_attention
         self.dropout = dropout
-        self.likelihood = likelihood
         self.add_relative_index = add_relative_index
 
         # initialize last batch size to check if new mask needs to be generated
@@ -484,7 +482,7 @@ class _TFTModule(nn.Module):
                                   skip=lstm_out if self.full_attention else lstm_out[:, encoder_length:])
 
         # generate output for n_targets and loss_size elements for loss evaluation
-        
+
         out = self.output_layer(out[:, encoder_length:] if self.full_attention else out)
         out = out.view(batch_size, self.output_chunk_length, self.n_targets, self.loss_size)
 
@@ -501,8 +499,137 @@ class _TFTModule(nn.Module):
 
         return out
 
+    def _produce_train_output(self, input_batch: Tuple):
+        return self(input_batch)
 
-class TFTModel(TorchParametricProbabilisticForecastingModel, MixedCovariatesTorchModel):
+    def predict(self, n, *args, **kwargs):
+        # since we have future covariates, the inference dataset for future input must be at least of length
+        # `output_chunk_length`. If not, we would have to step back which causes past input to be shorter than
+        # `input_chunk_length`.
+
+        if n >= self.output_chunk_length:
+            return super().predict(n, *args, **kwargs)
+        else:
+            return super().predict(self.output_chunk_length, *args, **kwargs)[:n]
+
+    @random_method
+    def _produce_predict_output(self, x):
+        if self.likelihood:
+            output = self(x)
+            return self.likelihood.sample(output)
+        else:
+            return self(x).squeeze(dim=-1)
+
+    def _get_batch_prediction(self, n: int, input_batch: Tuple, roll_size: int) -> torch.Tensor:
+        """
+        Feeds MixedCovariatesModel with input and output chunks of a MixedCovariatesSequentialDataset to farecast
+        the next `n` target values per target variable.
+
+        Parameters:
+        ----------
+        n
+            prediction length
+        input_batch
+            (past_target, past_covariates, historic_future_covariates, future_covariates, future_past_covariates)
+        roll_size
+            roll input arrays after every sequence by `roll_size`. Initially, `roll_size` is equivalent to
+            `self.output_chunk_length`
+        """
+        dim_component = 2
+        past_target, past_covariates, historic_future_covariates, future_covariates, future_past_covariates \
+            = input_batch
+
+        n_targets = past_target.shape[dim_component]
+        n_past_covs = past_covariates.shape[dim_component] if past_covariates is not None else 0
+        n_future_covs = future_covariates.shape[dim_component] if future_covariates is not None else 0
+
+        input_past = torch.cat(
+            [ds for ds in [past_target, past_covariates, historic_future_covariates] if ds is not None],
+            dim=dim_component
+        )
+
+        input_future = future_covariates[:, :roll_size, :] if future_covariates is not None else None
+
+        out = self._produce_predict_output(
+            x=(past_target, past_covariates, historic_future_covariates, input_future)
+        )[:, self.first_prediction_index:, :]
+
+        batch_prediction = [out[:, :roll_size, :]]
+        prediction_length = roll_size
+
+        while prediction_length < n:
+            # we want the last prediction to end exactly at `n` into the future.
+            # this means we may have to truncate the previous prediction and step
+            # back the roll size for the last chunk
+            if prediction_length + self.output_chunk_length > n:
+                spillover_prediction_length = prediction_length + self.output_chunk_length - n
+                roll_size -= spillover_prediction_length
+                prediction_length -= spillover_prediction_length
+                batch_prediction[-1] = batch_prediction[-1][:, :roll_size, :]
+
+            # ==========> PAST INPUT <==========
+            # roll over input series to contain latest target and covariate
+            input_past = torch.roll(input_past, -roll_size, 1)
+
+            # update target input to include next `roll_size` predictions
+            if self.input_chunk_length >= roll_size:
+                input_past[:, -roll_size:, :n_targets] = out[:, :roll_size, :]
+            else:
+                input_past[:, :, :n_targets] = out[:, -self.input_chunk_length:, :]
+
+            # set left and right boundaries for extracting future elements
+            if self.input_chunk_length >= roll_size:
+                left_past, right_past = prediction_length - roll_size, prediction_length
+            else:
+                left_past, right_past = prediction_length - self.input_chunk_length, prediction_length
+
+            # update past covariates to include next `roll_size` future past covariates elements
+            if n_past_covs and self.input_chunk_length >= roll_size:
+                input_past[:, -roll_size:, n_targets:n_targets + n_past_covs] = (
+                    future_past_covariates[:, left_past:right_past, :]
+                )
+            elif n_past_covs:
+                input_past[:, :, n_targets:n_targets + n_past_covs] = (
+                    future_past_covariates[:, left_past:right_past, :]
+                )
+
+            # update historic future covariates to include next `roll_size` future covariates elements
+            if n_future_covs and self.input_chunk_length >= roll_size:
+                input_past[:, -roll_size:, n_targets + n_past_covs:] = (
+                    future_covariates[:, left_past:right_past, :]
+                )
+            elif n_future_covs:
+                input_past[:, :, n_targets + n_past_covs:] = (
+                    future_covariates[:, left_past:right_past, :]
+                )
+
+            # ==========> FUTURE INPUT <==========
+            left_future, right_future = right_past, right_past + self.output_chunk_length
+            # update future covariates to include next `roll_size` future covariates elements
+            if n_future_covs:
+                input_future = future_covariates[:, left_future:right_future, :]
+
+            # convert back into separate datasets
+            input_past_target = input_past[:, :, :n_targets]
+            input_past_covs = input_past[:, :, n_targets:n_targets + n_past_covs] if n_past_covs else None
+            input_historic_future_covs = input_past[:, :, n_targets + n_past_covs:] if n_future_covs else None
+            input_future_covs = input_future if n_future_covs else None
+
+            # take only last part of the output sequence where needed
+            out = self._produce_predict_output(
+                x=(input_past_target, input_past_covs, input_historic_future_covs, input_future_covs)
+            )[:, self.first_prediction_index:, :]
+
+            batch_prediction.append(out)
+            prediction_length += self.output_chunk_length
+
+        # bring predictions into desired format and drop unnecessary values
+        batch_prediction = torch.cat(batch_prediction, dim=1)
+        batch_prediction = batch_prediction[:, :n, :]
+        return batch_prediction
+
+
+class TFTModel(MixedCovariatesTorchModel):
     @random_method
     def __init__(self,
                  input_chunk_length: int = 12,
@@ -516,7 +643,6 @@ class TFTModel(TorchParametricProbabilisticForecastingModel, MixedCovariatesTorc
                  add_relative_index: bool = False,
                  loss_fn: Optional[nn.Module] = None,
                  likelihood: Optional[Likelihood] = None,
-                 random_state: Optional[Union[int, RandomState]] = None,
                  **kwargs
                  ):
         """Temporal Fusion Transformers (TFT) for Interpretable Time Series Forecasting.
@@ -641,10 +767,15 @@ class TFTModel(TorchParametricProbabilisticForecastingModel, MixedCovariatesTorc
             # This is the default if no loss information is provided
             likelihood = QuantileRegression()
 
-        kwargs['loss_fn'] = loss_fn
-        kwargs['input_chunk_length'] = input_chunk_length
-        kwargs['output_chunk_length'] = output_chunk_length
-        super().__init__(likelihood=likelihood, **kwargs)
+        torch_model_params = self._extract_torch_model_params(input_chunk_length=input_chunk_length,
+                                                              output_chunk_length=output_chunk_length,
+                                                              **kwargs)
+        super().__init__(**torch_model_params)
+
+        # extract pytorch lightning module kwargs
+        self.pl_module_params = self._extract_pl_module_params(loss_fn=loss_fn,
+                                                               likelihood=likelihood,
+                                                               **kwargs)
 
         self.input_chunk_length = input_chunk_length
         self.output_chunk_length = output_chunk_length
@@ -757,8 +888,8 @@ class TFTModel(TorchParametricProbabilisticForecastingModel, MixedCovariatesTorc
             num_attention_heads=self.num_attention_heads,
             full_attention=self.full_attention,
             hidden_continuous_size=self.hidden_continuous_size,
-            likelihood=self.likelihood,
-            add_relative_index=self.add_relative_index
+            add_relative_index=self.add_relative_index,
+            **self.pl_module_params
         )
 
     def _build_train_dataset(self,
@@ -797,132 +928,3 @@ class TFTModel(TorchParametricProbabilisticForecastingModel, MixedCovariatesTorc
                                                n=n,
                                                input_chunk_length=self.input_chunk_length,
                                                output_chunk_length=self.output_chunk_length)
-
-    def _produce_train_output(self, input_batch: Tuple):
-        return self.model(input_batch)
-
-    def predict(self, n, *args, **kwargs):
-        # since we have future covariates, the inference dataset for future input must be at least of length
-        # `output_chunk_length`. If not, we would have to step back which causes past input to be shorter than
-        # `input_chunk_length`.
-
-        if n >= self.output_chunk_length:
-            return super().predict(n, *args, **kwargs)
-        else:
-            return super().predict(self.output_chunk_length, *args, **kwargs)[:n]
-
-    @random_method
-    def _produce_predict_output(self, x):
-        if self.likelihood:
-            output = self.model(x)
-            return self.likelihood.sample(output)
-        else:
-            return self.model(x).squeeze(dim=-1)
-
-    def _get_batch_prediction(self, n: int, input_batch: Tuple, roll_size: int) -> torch.Tensor:
-        """
-        Feeds MixedCovariatesModel with input and output chunks of a MixedCovariatesSequentialDataset to farecast
-        the next `n` target values per target variable.
-
-        Parameters:
-        ----------
-        n
-            prediction length
-        input_batch
-            (past_target, past_covariates, historic_future_covariates, future_covariates, future_past_covariates)
-        roll_size
-            roll input arrays after every sequence by `roll_size`. Initially, `roll_size` is equivalent to
-            `self.output_chunk_length`
-        """
-        dim_component = 2
-        past_target, past_covariates, historic_future_covariates, future_covariates, future_past_covariates \
-            = input_batch
-
-        n_targets = past_target.shape[dim_component]
-        n_past_covs = past_covariates.shape[dim_component] if past_covariates is not None else 0
-        n_future_covs = future_covariates.shape[dim_component] if future_covariates is not None else 0
-
-        input_past = torch.cat(
-            [ds for ds in [past_target, past_covariates, historic_future_covariates] if ds is not None],
-            dim=dim_component
-        )
-
-        input_future = future_covariates[:, :roll_size, :] if future_covariates is not None else None
-
-        out = self._produce_predict_output(
-            x=(past_target, past_covariates, historic_future_covariates, input_future)
-        )[:, self.first_prediction_index:, :]
-
-        batch_prediction = [out[:, :roll_size, :]]
-        prediction_length = roll_size
-
-        while prediction_length < n:
-            # we want the last prediction to end exactly at `n` into the future.
-            # this means we may have to truncate the previous prediction and step
-            # back the roll size for the last chunk
-            if prediction_length + self.output_chunk_length > n:
-                spillover_prediction_length = prediction_length + self.output_chunk_length - n
-                roll_size -= spillover_prediction_length
-                prediction_length -= spillover_prediction_length
-                batch_prediction[-1] = batch_prediction[-1][:, :roll_size, :]
-
-            # ==========> PAST INPUT <==========
-            # roll over input series to contain latest target and covariate
-            input_past = torch.roll(input_past, -roll_size, 1)
-
-            # update target input to include next `roll_size` predictions
-            if self.input_chunk_length >= roll_size:
-                input_past[:, -roll_size:, :n_targets] = out[:, :roll_size, :]
-            else:
-                input_past[:, :, :n_targets] = out[:, -self.input_chunk_length:, :]
-
-            # set left and right boundaries for extracting future elements
-            if self.input_chunk_length >= roll_size:
-                left_past, right_past = prediction_length - roll_size, prediction_length
-            else:
-                left_past, right_past = prediction_length - self.input_chunk_length, prediction_length
-
-            # update past covariates to include next `roll_size` future past covariates elements
-            if n_past_covs and self.input_chunk_length >= roll_size:
-                input_past[:, -roll_size:, n_targets:n_targets + n_past_covs] = (
-                    future_past_covariates[:, left_past:right_past, :]
-                )
-            elif n_past_covs:
-                input_past[:, :, n_targets:n_targets + n_past_covs] = (
-                    future_past_covariates[:, left_past:right_past, :]
-                )
-
-            # update historic future covariates to include next `roll_size` future covariates elements
-            if n_future_covs and self.input_chunk_length >= roll_size:
-                input_past[:, -roll_size:, n_targets + n_past_covs:] = (
-                    future_covariates[:, left_past:right_past, :]
-                )
-            elif n_future_covs:
-                input_past[:, :, n_targets + n_past_covs:] = (
-                    future_covariates[:, left_past:right_past, :]
-                )
-
-            # ==========> FUTURE INPUT <==========
-            left_future, right_future = right_past, right_past + self.output_chunk_length
-            # update future covariates to include next `roll_size` future covariates elements
-            if n_future_covs:
-                input_future = future_covariates[:, left_future:right_future, :]
-
-            # convert back into separate datasets
-            input_past_target = input_past[:, :, :n_targets]
-            input_past_covs = input_past[:, :, n_targets:n_targets + n_past_covs] if n_past_covs else None
-            input_historic_future_covs = input_past[:, :, n_targets + n_past_covs:] if n_future_covs else None
-            input_future_covs = input_future if n_future_covs else None
-
-            # take only last part of the output sequence where needed
-            out = self._produce_predict_output(
-                x=(input_past_target, input_past_covs, input_historic_future_covs, input_future_covs)
-            )[:, self.first_prediction_index:, :]
-
-            batch_prediction.append(out)
-            prediction_length += self.output_chunk_length
-
-        # bring predictions into desired format and drop unnecessary values
-        batch_prediction = torch.cat(batch_prediction, dim=1)
-        batch_prediction = batch_prediction[:, :n, :]
-        return batch_prediction
